Added Timings and Analysis
Added the timings and analysis of performance.
In the Excel Spreadsheet I have generated the average for each core from running the program in a batch script 10 times. I have chosen to use the average as It gives a more likely answer than taking one of the extremes e.g. min and max.